refacto: MentionnedUserInCommentRetriever::getMentionedUsers returns a collection of PFUser now
Part of story #38631 @ notifications in trackers

This refacto is needed because we need the user ID later while the
method returns a collection of username, it avoids having two different
method to retrieve the "same" thing.
Also, it removes the primitive (string) from the collection returned.

No functional change. Notification are still send when a new/edited comment is submitted (mass change NOT included)

Change-Id: Ib4ea02192b9ab84c43cc2dfb2c5af0af99d10d90