chore: request #40481 pnpm: 8.7.0 -> 8.15.9
Scripts executed during the lifecycle of packages installation are now
disabled. This is primarily to prevent an issue with the side effect
caches that is now enabled when `shared-workspace-lockfile` is set
to `false` [0]. While this could cause us some troubles by forcing us to
execute what we need manually it is likely a good thing. The pre and
post install hooks have been used in the past in situations where
deps were compromised as a way to run payloads.
Disabling the scripts execution also disable the recursive PATH update
in the repo which we compensate by adding our dev tools in the PATH
ourselves via the Nix shell.

Main changes:
https://github.com/pnpm/pnpm/releases/tag/v8.15.0
https://github.com/pnpm/pnpm/releases/tag/v8.14.0
https://github.com/pnpm/pnpm/releases/tag/v8.12.0
https://github.com/pnpm/pnpm/releases/tag/v8.11.0
https://github.com/pnpm/pnpm/releases/tag/v8.10.0
https://github.com/pnpm/pnpm/releases/tag/v8.9.0
https://github.com/pnpm/pnpm/releases/tag/v8.8.0

[0] https://github.com/pnpm/pnpm/commit/ef3609049f32e00718b073cd60fade03da30fbbf

Change-Id: I878ce0d972329f2a007cd2e29c403099ce01409c