fix: backlog vs kanban vs artidoc move icon
In backlog & kanban, cards can be moved to top|bottom.
In artidoc, section can be moved up|down.

Since the actions are not the same, we should have different icons.

Choice is made to take the same pattern than for pagination:
* double chevron for moving at the top/bottom
* simple chevron for moving one position

Positionning of icon is also fixed (there were some strange placement
with inconsistency between top/bottom).
Removes the spinner icon in backlog that was put only on down icon, and
was never displayed.

Part of story #40457: Modify flat document outline

Change-Id: I563b59a6dc67831bf6684c6220ae472297bbc0bf