refacto: Create a dedicated object for the post creation task
Part of story #38631 @ notifications in trackers

In later contribution we need to introduce the mentioned users, so
instead of adding yet another param in the PostCreationTask::execute
method, a dedicated object is created to handle some configuration. It will be easier to add new
params in the configuration class since we do not need to change 15488415 implemented classes anymore, there
is also less chance of breaking change and it also remove some primitive obsession.

No functional change expected

Note: Take the oportunity to remove the usage of Mockery in
AsynchronousActionsRunnerTest. (request #21775 Favor PHPUnit mock system
over Mockery - part 2)

Change-Id: I327cba8c2600825721dd17ac2f7bd858fcdd2476