feat: persist new order
When a section is moved (either with up/down arrows or with dnd), the
new order is saved.

If a pending section is moved, no call is made to the backend.
If an artifact section is moved, it is positioned relatively to the
next/previous artifact section.

Note: to keep things small, errors emitted by the backend are not handle
yet. It will be part of a next contribution, once we know how to manage
them from Design POV.

Part of story #40457: Modify flat document outline

Change-Id: I5ee084fc012c0658a4ffc5dfac31087a0c9d6cee