fix: pasted sentence can be rebuilt in wrong order
part of story #40114 Have better Cross References in Artidoc

When some content is pasted into the editor, the cross-references plugin
iterates over the Slice to be pasted. It finds the cross-references in the textual nodes, splits the text, adds marks and then rebuilds the sentences with the right order.

However, when a pasted sentence starts with a cross-reference AND
contains more than one cross-reference, then the sentence is rebuilt in
the wrong order. For instance, the sentence:

`art #123 and art #234 are referenced in this document.`

Will wrongly be rebuilt as:

`art #123 and  are referenced in this document.art #234`

How to test:
- paste `art #123 and art #234 are referenced in this document.` in your
  editor

--> The references have been rendered
--> The sentence is in the right order

Change-Id: I3a636e7293b9691ad56d5d0b05996168b3da9469