refactor: remove notion of service in artidoc domain
When we retrieve an Artidoc, we encapsulate it with the current service
so that the latter can be used to display header, footer, breadcrumbs,
or retrieve by transitivity the current project.

The project or service, for now, do not need to be in artidoc domain
because artidoc is about composing a document of sections.

This contribution proposes to introduce the concept of "context" for an
artidoc: the adapter inject its context during the artidoc retrieval, so
that it can be used later when needed. This is the same design approach
that is used with our middlewares: they check if the project is valid,
or if the user can access the resource and inject them as attributes of
the request to be used in the controllers.

This relies on the usage of instanceof (just like Request attributes),
but at least the domain does not define something that it should not
know.

No functional changes.

Part of story #40457: Modify flat document outline

Change-Id: I2a3c71624e4348495a3664a9a84c3e344e0d990b