fix: dragged section should be hidden in the list
part of story #40457: Modify flat document outline

This fix consists in removing a check in @tuleap/drag-and-drop.

The check, when it was successful, allowed the dragged item to be hidden
in its container. Sadly, it was failing in our use case. Removing it
completely seems to have no impact on the existing apps using the lib
(taskboard, program planning).

We can always revert this commit if we notice something broken because
of it.

How to test:
- Drag a section in the TOC to reorder it
--> The original section in the list should be hidden

Change-Id: I0723bd6d7ec2f6bd5721dff4cfcc8c952fab336d