refactor: Move ArtidocRetriever in the domain
The business logic to retrieve an artidoc with context (build the
document, check the permissions, …) is now part of the domain. The
implementation is moved in the adapter (to interact with the docman or
with the Tuleap services).

In order to not pull too much hair off my head, a tradeoff is taken
about the management of the user: it is injected in the constructor of
the adapter. Usually parameters that depends on the current request
should be passed at the method level and we should keep constructor
parameters for infrastructure dependencies.
However this choice is interesting here since it removes completely the
need to define something for the user in the domain (no need to rely on
user "identifier" which −forces us to re-retrieve the user afterwards
and− does not bring much value in the domain).
We can still change this decision later when really needed.

Suggest entry point for the review:
\Tuleap\Artidoc\Domain\Document\ArtidocWithContextRetriever::retrieveArtidoc

No functional changes.

Part of story #40457: Modify flat document outline

Change-Id: Ic7ce7e8efe706c21bd32e541b93c5c26d6b4aac2