Deptrac should fail on CI on tests/deptrac/core_on_plugins_skip_uncovered.yml
part of request #40476 Improve deptrac

CI should fail when deptrac rules are not respected

How to test
After merge of gerrit #32579 and gerrit #32562
CI should be happy, before it should fail

Why?
It seems that havin `tests` in excluded_files enable in weird behaviour on CI, file to analyse are never found

Trace of progress and verbose execution:
```
Start to create an AstMap for 0 Files.

    0 [░░░░░░░░░░░░░░░░░░░░░░░░░░░░] < 1 sec
```

Note:
I'm not sure its the best option, but I took the regexp for excluding tests file directly on deptrac documentation
https://qossmic.github.io/deptrac/configuration/

Change-Id: Iab68413ed48623f5f408e60dac479df16c327507