chore: request #40478 Introduce internal vue2-dompurify-html lib
This is a limited version of vue-dompurify-html without a compat layer
to support both Vue 2 and 3 to be able to continue support as needed.

On purpose, only the types of the latest version of Vue 2 are exposed
which can cause some type mismatch with older Vue versions. The type of
Vue 2.7.16 are closer to the ones available in Vue 3 so it is one more
reason to migration :)

No functional changes.

Change-Id: Id76578188940d9afb49f2be0b699b35a2d217a20