fix: A before C should move ABC to BAC
Part of story #40457: Modify flat document outline

What has been fixed:

Given three section A, B and C.
When A is moved before C
Then A is in fact moved after C

This issue happens when the moved section is before another section
further in the list because when the moved section is removed from the
list, the indexes of the sections are mutated. Once we try to put back
the section at the right place, the sibling section index is no longer
valid: the section is moved after its sibling, and not before it.

Before reordering:

sections = [A, B, C]
indexes  =  0  1  2

The target next sibling index is 2
The moved section index is 0

We remove the section from the list:
sections = [B, C]
indexes  =  0  1

We add back the moved section at the sibling's former index which was 2:
sections = [B, C, A]
indexes  =  0  1  2

KO

This behavior does not happen when the moved section is after the
sibling section in the list because the sibling section index won't be
mutated following the moved section removal from the list. Example when
reordering C before A:

Before reordering:

sections = [A, B, C]
indexes  =  0  1  2

The target next sibling index is 0
The moved section index is 2

We remove the section from the list:
sections = [A, B]
indexes  =  0  1

We add back the moved section at the sibling's former index (still 0):
sections = [C, A, B]
indexes  =  0  1  2

OK

How to test:
- Have three sections A, B and C
- Move A before C
--> A has been moved before C

Change-Id: I67cfe2f57549967b29d92d225d6d3b01741192f6