feat: open toolbar's image menu when pressing ctrl+k in the editor
part of story #38627: Choice of a new Rich Text Editor

This contribution introduces a new shortcut to the editor. This shortcut
allows the user to open the toolbar's image menu by pressing the Ctrl+k
keys.

[bonus] When the link, image, edit link/ref menus open, their first
inputs in their forms are now automatically focused. The code achieving
this is in the lib @tuleap/tlp-popovers because we need to wait until
popovers are shown before being able to focus what we need.

[bonus 2] When the link or image menu is closed, the editor gets its
cursor and focus back.

How to test:
- Click in an editor
- Press Ctrl + k
--> The image menu is open
--> The "Image source" field is focused

- Insert a new image then submit OR cancel
--> The editor gets its cursor and focus back

- Click on the link button
--> Its menu opens
--> The "Link" field is focused

- Close it
--> The editor gets its cursor and focus back

- Open the edition form of a link
--> Its "Link" field is focused

- Open the edition for of a cross-reference
--> Its "Reference" field is focused

Change-Id: Ib26c36f734833d46562e41023b8cc0b589b5cce7