feat: warning when tracker has no notifications
part of story #38631 @ notifications in trackers

How to test:
- As tracker admin, go to th tracker's administration > Notifications
and disable all notifications.
- When you edit an artifact of that tracker, there is now a warning
alert below the textarea for new comments. It warns that at mentions
will not send any email.

Why?
Template rendering is moved to Field_Text and Edit classes. This way, in
the case of the new follow-up comment, we can include it in a bigger
dedicated template, but leave it as-is for text fields.

Change-Id: I873f25396306259b71ee4c075037550be83ebc81