fix: Purified content is not rendered anymore
part of request #40478 Introduce internal vue2-dompurify-html lib

How to test:
- Without this patch, delete the frontend-assets/ and node_modules/
folders in the affected apps (project-labeled-items, program management,
etc.)
- Rebuild (make post-checkout)
- v-dompurify-html in those apps should be ineffective: nothing is
rendered:
- In labeled items widget, nothing is rendered for the pull-request
icon and title.
- In program management, when the program is not configured, the
link to administration is not rendered (which was caught by nightly
end-to-end tests).
- In trackers, when you browse a Test Definition artifact, and you
hit "Delete" button on a test step or you click on "Reorder steps"
button, the steps contents is not rendered.
- In trackers, if you setup a tracker without any static selectbox
(for example, only Title and Description), and you go to the tracker
administration > Workflow > Transition rules, the workflow app should
show a warning with a link to the fields administration, but it shows an
empty alert box.

- With this patch, after rebuild, in all the cases above, things are
correctly rendered, links and icons are visible.

Why?
With the upgrade to pnpm v8.15.9, lifecycle scripts are not run anymore,
but vue-dompurify-html requires the lifecycle script of vue-demi to run
with Vue 2. Instead of working around that, we chose to fork the Vue 2
version of the library and use this. The remaining Vue 2 apps that were
still using vue-dompurify-html are now switched to the Vue 2 version.

Follow-up of 69c8917b2011a7ee7134a5d82ee0e05d9e74f25c
Issue revealed by bdb65ea417bbbf9b935046496c8b8f8e658d6d6f

Change-Id: Ifff55747ab208ae8140bd88ae19b23d59320849b