refactor: stop relying on representation to check permissions
The creation of new sections was relying on transforming the given
section to its representation to make sure that current user was allowed
to read it before adding it to the document. This situation was fine
during the early days of the prototype, but since our baby is growing,
we must change this stinky diaper, so that layers are not mixed.

No functional changes.

Part of story #40457: Modify flat document outline

Change-Id: I7b3b65cea5ef399b24d0ed0d3a9cbd7e1e396d10