feat: info and warning messages in comment edition
part of story #38631 @ notifications in trackers

How to test:
- As tracker admin, go to the tracker's administration > Notifications
and disable all notifications.
- When you edit an artifact of that tracker, when you edit a comment,
there is now a warning alert below the textarea. It warns that at
mentions will not send any email.
- If you enable all notifications, the alert is now an info box.

Why?
I chose to go with an attribute only when for the disabled case. I
expect that there are more trackers _with_ notifications than without,
this avoids adding an attribute for the most common case.

Change-Id: Ie5ebf437bf8cf2a78c04252b288ec43174f5f1bf