feat: init atwho autocompleter in comment edition
part of story #38631 @ notifications in tracker

How to test:
- When you edit a comment with any format (HTML, Text, Markdown), or
when you change the format, and you type "@" + 3 letters, an
autocompleter will suggest a list of valid usernames with their avatars.

Note:
For HTML format, it was already functional as init is handled by
rte-creator lib. For new follow-up comment, the init is done as a
side-effect of importing the @tuleap/mention lib. We should change that,
as it is confusing.

Change-Id: I49ee4b88f807ae7cb3cdfc05a8cff19311f89fb9