refactor: remove side-effect from mention lib
part of story #41090 @ notification in tracker mass-change
& artifact modal

No functional change expected:
- when you edit an artifact, when you write a new comment, the
autocompleter appears (when you type "@" + 3 letters).
- when you edit a comment, the autocompleter appears.
- when you browse a release cardwall, and you edit a story or the
release, in the modal v2, when you write a new comment, the
autocompleter appears.
- no change for usages that did not rely on user-mention class: the
autocompleter still appears when you add or edit a Static text field in
Tracker administration > Fields, etc.

Why?
Consumers of this lib should be responsible of making the init, not the
lib itself. It removes some magic, which is good, as it makes it
possible to understand what's happening. It also removes the "magic
string" of the ".user-mention" CSS class that was used as a selector to
enable the autocompleter.
Having the rte-creator lib deal with all cases of mention removes also
some of the magic. It now sets up mentions both in the HTML case
(onEditorDataReady) and in the Text and Markdown cases (initMentions on
the textarea).
I took the opportunity to switch from a class to a class-like function.
It removes a bunch of `this` and `new`.

Change-Id: Ie60aecdc1959fc13973b3127f01225acb21a5017