fix: keep focus on up/down button
In order to ease the move of sections with the keyboard, we should keep
the focus on the button so that the same section can be moved up or down
easily without having to navigate back to it.

The regression has been introduced by
git #tuleap/stable/e23e5f86ef95dcebe780e1a51f407137017fe246 by hiding
buttons until the save is done by the backend so that user do not start
to reorder things that are not saved yet. However, using `display: none`
(or `visibility: hidden`) looses the focus on the button. Instead the
opacity is now set to 0.1 to appear somewhat "disabled" (and
pointer-events are disabled so that user doesn't interact with those
elements).

To test: navigate with the keyboard until you reach the arrow button
down of the first section. Hit enter, the section order is saved, and
you still have the focus on the same button (no need to navigate to it).

Part of story #40457: Modify flat document outline

Change-Id: Ic25b798abeec68dffb1aaf92c4d488622315aa53