fix: markdown like in the title
Part of story #38627: Choice of a new Rich Text Editor

what has been fixed?
When writing characters that correspond to inputRules in the title, the characters were deleted. For example, `*word*` became `word`.

The tests have been modified to test this behavior and also test each rule.

How to test:
-- Write *word* in the title
-> Nothing should happen and *word* should still be written

Change-Id: I9b3c7aac3abf6a9308812e68fdf7dc1a6e0a18a7