refactor: move SectionCreator to domain
No functional changes

Implementation details
----------------------

The steps to create a section in an artidoc are:
1. Retrieve the document and make sure that user can write
2. Retrieve required artifact information of the new section and make
   sure that user can read it
3. Save section at the given position
4. Return the section representation based on a) the section identifier
   given by step 3. and b) the artifact information gathered at step 2.

Step 1. is already divided in Port/Adapter
Step 2. must be part of the Adapter layer because we need to
target the tracker internal API to retrieve title or description field
of the artifact and to ensure that current user can read it.
Step 3. can be easily moved to Port/Adapter (interface -> Domain)
Step 4. must be part of the "Adapter" layer since it is at the REST API
level.

Current challenge: interwind Domain/Adapter 2 way round trips: Domain
cannot pass back retrieved required artifact information to the Adapter
layer in case of success because those data are for the Adapter only.

In order to bypass this, Adapter data stays in Adapter layer thanks to a
"collector". Domain tells to the Adapter "make sure that you collect
required information, if it is ok then I will create the section".
Collected information are then used by the REST API layer to build and
send the section representation.

Part of story #40457: Modify flat document outline

Change-Id: I3dbfce54701c09431f4ae2e28326b9fe0f947112