refactor: comment format → enum
part of story #41090 @ notification in tracker mass-change
& artifact modal

No functional change when creating changesets.

Why?
CommentFormatIdentifier was probably introduced before PHP supported
enums. It is a good candidate because it needs a conversion to string
for the database and legacy methods, and there can only be 3 cases.
Besides, enums are readonly and cannot be extended by default.

Change-Id: I0f3f82800d9f7c3a27eeba9313b98ea5e049fb78