fix: artidoc e2e tests
Part of story #40457: Modify flat document outline

Drag-and-drop tests to reorder sections failed on the nightly.
I was unable to reproduce the error, either on the interface or in headless mode.
We decided to remove them and keep only the tests to reorder the sections with the up and down arrows.

Note that drag & drop didn't work with `slow 4G` but worked fine otherwise. It seems that `dragstart`, `drop`... events were not triggering with a slow connection but I don't know why or how to solve this.

I took the opportunity to add @postSections waits to avoid test errors in the event of network latency.

How to test:
-- Remove `cy.wait(["@postSections"]);` on line 103. Run the tests with `slow 4G` (console>Network>No throttling -> slow 4G or less) on Cypress.
-> The tests should fail because the page is reloaded before the last sections have been saved.

Change-Id: I859d478bc23617cad09fcb953eb47027c7fce301