chore: request #41093 move git backend assets in a sub-directory
How to test:
No functional change. If you commit a file with a special unicode
character in it (for example \u202a), when you browse the Blame for this
file or the diff for your commit, there should still be a message at the
top warning you that the character could hide something.

Why?
Having a package.json at plugins/git/ interferes with Turbo's caching
behaviour. Every time there is any change in the Git plugin, it will
want to rebuild the backend assets. By moving it in a sub-directory, it
will now stay cached, unless it changes. We also move the e2e
dependencies to their own folder.

Change-Id: I402dc00344abab7c9c704628cc3f7f301e736f42