refacor: Remove PUT /artidoc/:id/sections
This route was designed during the early stage of artidoc plugin in
order to be able to test the rendering of a document.

Now that user can create section or add existing section, this route is
not useful anymore.

No functional changes, since it was not used.

Part of story #41089: Prepare Artidoc backend for Structure

Change-Id: I2572f99b78c981cbdcac6742ca16e9a609b1daf6