refactor: replace translate directives by gettext
part of request #41095 Migrate plugin-program-management-iterations to
Vue 3

No functional change expected.

Why?
The v-translate and translate Vue directives are deprecated and removed
in Vue 3. This is mandatory for the conversion.
The snapshot test broke, so it was removed. It is more effectively
replaced by limited assertions that will not break when we change a CSS
class or a translated string.

Change-Id: I75df6ad209217b803f24b5b19e9c9c45746c6af1