fix: Absolute url for link to user when @mention
Part of story #41090 @ notification in tracker mass-change & artifact modal
Fixes request #14674 Invalid URL in email notifications for @mentions

When @mention is used in an artifact comment, in the email the link was
relative.

Note: As all mention to link transformation are at the same place, there
is no difference if the link is in Tuleap or in a mail. So all mention
link are now absolute. It should not break anything as the same
treatment is done for references.

*Testing:*

Mention a user in a follow-up comment, in the email the link should
works and bring you back to Tuleap

Change-Id: Iecae79d7a1093e02f0af0bf1755606080dfbf8c5