feat: Add notification status in Tracker REST representation
Part of story #41090 @ notification in tracker mass-change & artifact modal

We need to know if tracker notifications are enabled or not in artifact
modal. For that we add a new key in Tracker REST representation:
"notifications". It's an object because we don't know as of today if it
will need to contains more data.

*Testing:*

Fetch /api/trackers/{id}. The json should contains:

    "notifications": {
        "enabled": <boolean>
    }

The boolean should correspond to your tracker notification setup

Change-Id: Ied80de2ab1059581b73a57641b66409c73a089a9