refactor: split API client
part of story #41090 @ notification in tracker mass-change & artifact
modal

How to test:
No functional change in the artifact modal link field. e2e tests in the
agiledashboard and kanban plugins should pass.

Why?
We need to add another function in TuleapAPIClient to retrieve the
existing artifact and its tracker structure, so that we know whether
notifications are disabled. It will kill another part of the
"rest-service" from the Angular.js days.
The TuleapAPIClient was becoming too big (> 600 lines of unit tests).
Since it was designed to implement many single-function interfaces, it
was easy to split it in three: one client for the link field in general,
one client for the creation of artifacts in the link field, and one for
the rest.
I also grouped related files in a creation/ folder in the REST adapter.

Change-Id: I60e9ae6460ba429522d11b4fc6f1cfff9f3f980f