refactor: Use API client to retrieve current artifact & tracker
part of story #41090 @ notification in tracker mass-change & artifact
modal

How to test:
No functional change in the artifact modal. e2e tests in the
agiledashboard and kanban plugins should pass.

Why?
We need to fetch another bit of info from the tracker structure, so that
we know whether notifications are disabled. To do so, we first kill
another part of the "rest-service" from the Angular.js days.
We now have an "Initialization" API client because TuleapAPIClient needs
the current project ID in its constructor, and we don't have it yet
(it's on the tracker).
CurrentArtifactIdentifierProxy is replaced by a plain "constructor"
fromId() in the domain. The "modal Service" is now responsible for
building it once, and passing it to the rest. It reduces the number of
classes to deal with, and anyway the Domain interface + Proxy pattern
could not guarantee that we don't create such an identifier from thin
air (it's even less strong in TypeScript than in PHP).
I also removed remaining things in the controller that were either
duplicate or dead.
Removing $q seems to have no effect on the code creating the modal,
edition model tests in the modal Service should be simpler.

Change-Id: Ide2315841363457252783d88e75a8f65d2b75ac7