fix: add new section dropdown behind the section title
Part of story #41105: Have freetext section in artidoc

What has been fixed:
When you click on "add new section" button and move the cursor outside of the drop-down menu (top or left), the bottom of the drop-down menu disappears behind the title instead of slowly disappearing like the rest of the drop-down menu.
It's the same when you move the cursor back into the drop-down menu (less visible).

How to test:
-- Click on "add new section", go to the drop-down menu and move the cursor outside (by the top)
-> The entire drop-down menu should slowly disappear

-- Move the cursor back in drop-down menu
-> The entire drop-down menu should appear

Change-Id: I962fd65fa4b0f82c9dd0d80f3b1bc8edc0132b3d