feat: GET /artidoc_sections/:id for freetext
If you hack the database to introduce a freetext section in a document,
then in the API explorer you can get the content of this section by
using the section uuid you used to forge the data.

No functional change for artifact section.

Part of story #41105: Have freetext section in artidoc

Change-Id: Ib2dfb8a1e80cb8bbed82388e54000e59d1815ebb