refactor: Extract mentioned users retriever to core
Part of story #38643 @ notifications in pull requests

For pullrequest we need the exact same class to retrieve users in
comments. But the current one is in tracker plugin and so out of range
for pr. By extracting it in the core it is available for tracker and for
pr (and the rest of the world).

*No functional changes expected*

Mentioned users notifications in trackers should continue to work as
before

Change-Id: I396486685cb1727782db3486014ef8b589af608f