refactor: extract section reordering from store
story #41105 Have freetext section in artidoc

Why?

The management of the sections order has been put in the SectionsStore
because this store holds the sections, and it was convenient to put
the code inside it to manipulate them.

However, the stores in artidoc are not that great because all the logic
ends up in them, even stuff related to the UI. They size tend to
increase through time, and they are slowly becoming god objects.

Moving sections should not be part of any store, because it is a feature
belonging to the TOC component. There is not that much of a reason to
share the feature with the whole application. Instead, we can have a
class-like object taking a ref containing the sections to achieve the
exact same thing.

Note:

For now, I've only extracted the code into one single object. I've tried
to split it in different parts, but I felt like it wasn't the right time
to do it. Let's reconsider this second refactoring once we'll know how
to reorder nested sections.

-- no functional change expected

Change-Id: Ia894bd4337e32fdacf0590b80cfd7abb7d706f0b