feat: init atwho autocompleter in modal
part of story #41090 @ notification in tracker mass-change & artifact
modal

How to test:
- When you edit a comment with any format (HTML, Text, Markdown), or
when you change the format, and you type "@" + 3 letters, an
autocompleter will suggest a list of valid usernames with their avatars.
- No functional change on autocompleter in artifact edit view or mass
change.

Why?
Part of the CSS was in FlamingParrot, it has been moved to the
@tuleap/mention lib itself.
Now that the lib does not init itself (since
069470192a0aa07702eb0704532f0eddf749e0f3), we can remove the forced
content-editable attribute that was added on ckeditor's document.
Also cleaned-up RichTextEditor and impacted tests while I was there.

Change-Id: I210913b3c9abb348d2de3dc2e2b9c61e868c45d8