refactor: get locale for image upload from controller
part of story #41090 @ notification in tracker mass-change & artifact
modal

No functional change in image upload in new comment or text fields in
the artifact modal

Why?
We should retrieve the locale from document body only once, and then
pass it to the rest of the elements. We should not treat document body
as a global store, but treat it the same way as a mount point.

Change-Id: I49b0721b81eabd869253492a98fb327f53d5c3a9