feat: Notify mentioned users in PR global comments
Part of story #38643 @ notifications in pull requests

User mentioned with @user in global PR comments are now notified

*Testing:*

In a git pull request, create a new comment and mention a user in it. It
should receive a notification.

Update this comment (or another one) and add a mention, the user should
be notified too.

Change-Id: Ib56157fb9e9c70a8e85c0e52fd367bd4b0591bbc