feat: replace alert by text info/warning for mention message
part of story #41090 @ notification in tracker mass-change & artifact
modal

How to test:
- When you edit an artifact, the at mention info and warning messages
are not alerts but text info/warning without icon, as per the mockup.
- The info message about image drag/drop is now also a text-info in the
artifact edit view, in the Step definition field from Test Management,
in the modal v2 in Cardwall and in the artifact modal (v3) in Kanban.

Why?
UploadImageFormFactory is changed to a class-like function (no need for
"this" and "new"). HelpBlockFactory too.
Removed the wrapping div for the image upload help message, a single "p"
is enough.
I changed the translated string to remove an abbreviation ("'n") and fix
an English mistake. Translations stay the same.

Change-Id: I7c493012f72f7996e7536276e79dcf711c208a11