feat: PUT artidoc_sections/:id
Freetext section content can now be updated via the REST API.

If you try to update an artifact content with this route you will get an
error. This will probably handled later instead of relying on
/artifacts/:id route => we will have only one entry point to update a
section, and the artifact section adapter will delegate the update to
the tracker layer. I suspect this will be mandatory when we will
implement artidoc history.

To test you need a freetext section:
* Create a dedicated document
* Create two artifact section A & B
* In the database, in `plugin_artidoc_document` for section A, remove
  the artifact id, and copy the section identifier in the freetext id
  column.
  Use the same freetext id to create an entry in
  `plugin_artidoc_section_freetext`
* Call GET artidoc_sections/:id with the section identifier
  => you should get the freetext content
* Call PUT artidoc_sections/:id with a new content
  => you should get 200
* Call GET artidoc_sections/:id
  => you should get the new freetext content
* Call PUT artidoc_sections/:id with the section id being the section B
  => you should get 400

Part of story #41105: Have freetext section in artidoc

Change-Id: Ifa0254898a2c859e4cfdcb6d3ea17b5ed4bc2ca6