feat(forge-config): list how to modify config keys
`tuleap config-list` now explain where a given configuration variable
can be modified.

Until now we had the binary database/file choice where database was the
only described way to modify a variable. As config-list aims to become
the central tool to identify and manage configuration variables, it
should be more precise information on how variables can be modified:
database, files or no modifications permitted.

I choose to keep the name of the original interface for the attribute
ConfigCannotBeModifiedYet to avoid a fat commit. It's basically an alias
for "modified in local.inc".

I also choose to make database variables "ConfigCannotBeModifiedYet" so
we can advertise where the variables can be modified.

Finally I took the opportunity to make database password a Secret config
key. The attribute didn't exist when we did the original implementation.

Part of: request #40460 Have a single source of truth for configuration variables

Change-Id: Id8ba113fa81045443592177bbec5ba2ed5788753