feat: Notify mentioned users when creating a PR
Part of story #38643 @ notifications in pull requests

When creating a PR its description is filled with the commit
description. If it contains some @mention, targeted users are now
notified about the PR creation

*Testing:*

Create a pull request from a commit with some mention in its
description. Concerned users should receive a mail

Change-Id: Ib4812babc628a59f7402b6ece4c1c7a2d7389102