fix: Cannot add an existing section when no tracker is configured
Fixes request #41117 Cannot add an existing section when no tracker is configured

Note 1:

The configuration store does not execute the onSuccessCallback anymore since we need to perform actions when the modal is closed AND the save was a success.

For the record, the modal works like this:
1. Select a tracker
2. Submit
3. If failed: feedback error, cannot do anything but canceling, else success feedback, click "Close" to continue the edition of the current artidoc.

We now need to pause the opening of a second modal until a condition is fulfilled (tracker configured successfully). If we keep letting the configuration store execute the callback, we'll have the "add existing modal" displayed on top of the configuration modal as soon the call to PUT /artidoc/{id}/configuration returns 200, which is not acceptable.

Hence, we need to wait until the user has closed the modal to execute the callback:
- opening the "add existing section" modal when the user needs to add an existing section
- pushing a pending artifact section when the user wants to create a new section.

Note 2:

The modal dismissal when its backdrop is clicked has been disabled because we need to make sure we do the proper actions when the user closes the modal AND the save has been successful:
- Using the cross in the header
- Using the cancel button
- Using the close button

Clicking the backdrop is handled by @tuleap/tlp-modal which does not allow us to do anything in this case, which is why I prefer disable it, especially since this modal has 3 different states (chose tracker, success, fail).

Alternatively we could chose to not disable it and add an event listener on TLP_MODAL_HIDDEN, but since the configuration modal has a specific UX flow with different states, it seems better to prevent it to be mistakenly closed if the user clicks outside of it.

How to test:
- Have an artidoc
- Using its id, delete its configuration from the
  plugin_artidoc_configuration_tracker table
- Go to the artidoc
- Click on the ( + ) button
- Click on [Add existing section]
--> The configuration modal should open

- Select a tracker, then submit
--> Success feedback

- Close the modal using the [Close] button
--> The configuration modal is closed
--> The "add existing section" modal opens
--> You are able to search artifacts to add to your artidoc

Change-Id: I20a702b7d4f81406552b7faaccf33e3508ea9d41